Shiny projects not on git yet

+++ 1-tidy/1-L2M-bkref.R
@@ -467,5 +467,7 @@ l2m_shiny <-
          temp_date = as.Date(paste(temp_year, month(date),
                                    day(date), sep = "/"), "%Y/%m/%d"))
 
-write_csv(l2m_shiny, "3-shiny/ref-shiny/L2M.csv")
-write_rds(l2m_shiny, "3-shiny/ref-shiny/L2M.rds")
+if (file.exists("3-shiny/ref-shiny/L2M.csv")) {
+  write_csv(l2m_shiny, "3-shiny/ref-shiny/L2M.csv")
+  write_rds(l2m_shiny, "3-shiny/ref-shiny/L2M.rds")
+}
